+++ Packages/eu.netherlands3d.twin-features/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- IConfiguration interface to declare configuration per feature including SimpleJSON mapping and QueryParameter mapping
+- IQueryStringMapper to help populate feature configuration from a URI string
+- ISimpleJsonMapper to help populate feature configuration from a JSON object
+
 ## [1.0.0] - 2023-06-13
+
 ### Changes
+
 - Initial release